Create README.md
Have your phone ever accidentally dropped in water? It is a frustrating experience that can damage important parts of your phone like the screen, buttons, microphones, camera, and especially the speakers. Sometimes, the water gets into other parts of the mobile except the speaker, so it becomes challenging to remove water from it without having a technician show up. But in the case of the speaker, you can remove this water by (Fix MY Speaker) with a sound of a particular frequency of 165 HZ.